+++ test.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-sudo python detux.py --sample _testing/example_binary1 --report test_out.json --timeout 30
+sudo python detux.py --sample _testing/example_binary1 --timeout 30
