+++ wintest.sh
@@ -2,5 +2,4 @@
 
 sudo python detux.py \
     --sample _testing/5DA8C98136D98DFEC4716EDD79C7145F \
-    --report test_out.json \
     --timeout 30
